Latest changes: adding file analysis.py to test program outputs on test samples set new parameters on video pipeline for better quality
